Rename .compat.pyam.computations to .operator
Adjust usage and references.
