+++ genno/compat/pyam/operator.py
